Fixed Bug #1
In this commit, the bug 1 was fixed by putting an "if, else" statement inside the handleDrop function, that checks if the drop zone has any child nodes already in it (this.childNodes.length > 0). If it is already filled, the function returns and doesn't allow the drop. If the drop zone is empty, it allows the drop.